Fixed app work on devices with <28 API; Added base currency amount criteria to rates observation

+++ data/src/main/kotlin/com/vbta/currenciesta/data/source/remote/model/CurrenciesRateApi.kt
@@ -11,5 +11,5 @@ data class CurrenciesRateApi(
     @Json(name = "date")
     val date: Date,
     @Json(name = "rates")
-    val rates: Map<String, Float>
+    val rates: Map<String, Double>
 )
